extract clipping functions to new clip module
