+++ src/deepqmc/conf/task/train_ferminet.yaml
@@ -20,7 +20,7 @@ pretrain_kwargs:
     basis: sto-6g
 fit_kwargs:
   clip_mask_fn:
-    _target_: deepqmc.fit.median_clip_and_mask
+    _target_: deepqmc.clip.median_clip_and_mask
     _partial_: true
     clip_width: 5.0
     median_center: false
