+++ src/deepqmc/conf/task/train_saml.yaml
@@ -19,7 +19,7 @@ pretrain_kwargs:
     basis: 6-311G
 fit_kwargs:
   clip_mask_fn:
-    _target_: deepqmc.fit.median_clip_and_mask
+    _target_: deepqmc.clip.median_clip_and_mask
     _partial_: true
     clip_width: 5.0
     median_center: false
